NEW FEATURE: score pop-up message
Problem
It is needed to show how many points has
been added to the score after gems have
been destroyed.

Solution
The score pop-up text message was added.
According to the gems position and the
number of the gems the message changes
its location, color and size.

Note
The position is not very accurate,
can be changed later.